+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Cookie Management for http client to support clustered environments with cookie based sticky sessions
 - Raise an exception, if authenticator is defined for a basic-flow execution
+- Support for managing `Client Scope Mappings` like other resources by configuring `import.managed.client-scope-mapping=<full|no-delete>`.
 
 ### Changes